Refactor: Move utility functions from utils.ts to helpers/utils.ts
- Updated import paths across multiple files to reflect the new location of utility functions.
- Removed the old utils.ts file and its associated stub for testing.
- Adjusted tests to import from the new helpers/utils.ts file.
