+++ src/commands/command.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import * as utils from '../utils';
+import * as utils from '../helpers/utils';
 import { exec, spawn } from 'child_process';
 import { outputChannel } from '../extension';
 import * as fs from 'fs';
