+++ src/commands/fcli.ts
@@ -3,7 +3,7 @@ import * as fs from "fs";
 import * as YAML from "yaml";
 import { execCommandInTerminal } from "./command";
 import { ClabLabTreeNode } from "../treeView/common";
-import { getSudo } from "../utils";
+import { getSudo } from "../helpers/utils";
 
 function buildNetworkFromYaml(topoPath: string): string {
     try {
